Add files via upload
Created the Chess Engine file which contains a class responsible for the base implementation of the pygame gui. Essentially the board is set up to appear on the screen, and the user can either hit the close symbol on the gui or press ctrl-q to exit the program. The skeleton for introducing interaction between the user and the board has also been placed in the file. I also added a class to chess.py for the board state and rules of Chess to reside. In addition, main is set-up to launch the program as well.